requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.5.0
altair==5.0.1
astunparse==1.6.3
attrs==24.2.0
av==10.0.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==4.2.4
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
fonttools==4.38.0
gast==0.3.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
idna==3.10
ifaddr==0.2.0
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
jinja2==3.1.6
jsonschema==4.17.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b1
pyee==10.0.2
pygments==2.17.2
pylibsrtp==0.8.0
Pympler==1.1
pyopenssl==25.1.0
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==308
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9.1
scipy==1.4.1
six==1.17.0
smmap==5.0.2
streamlit==1.23.1
streamlit-webrtc==0.44.7
tenacity==8.2.3
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.2
typing-extensions==4.7.1
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
